WIP - Transparently proxy incoming URI without decoding
Previously we decoding the incoming URI to a net.URI object and then
encoding it back to a URI object to make the outgoing request. On the way
through this process, due to the way golang handles encoding and
decoding, the URLs sometimes silently change. For example certain
characters such as ! would get percent-encoded before making it to the
backend. This commit uses RequestURI instead so that we use exactly the
incoming request path in the outgoing request.

[#77014252]

Signed-off-by: Michael Fraenkel <fraenkel@us.ibm.com>